feat: Add @nexusai/shared package to API dependencies
This commit adds the "@nexusai/shared" package to the dependencies in the "package.json" file of the API. The package is added with the version "1.0.0". This change is necessary to support the country validation feature in the phone number utils and improve data integrity when creating agents with valid phone numbers.