refactor: Update utils.ts to include logout function
This commit adds a new logout function to the utils.ts file. The logout function removes the "token" and "user" items from the localStorage and redirects the user to the logout endpoint of the API. This change improves the security of the application by ensuring that user data is properly cleared when logging out.