refactor: Limit the number of scraped links to 8
This commit updates the `scrapeLinksFromWebpage` function in `scrapper.ts` to limit the number of scraped links to 8. By introducing the `MAX_LINKS` constant and using it to slice the `_links` array, we ensure that only the first 8 unique links are returned. This change improves the efficiency of the scraping process and prevents an excessive number of links from being processed.