```text refactor: Add Pause icon and retrainKbData/deleteKnowledgeBase functions
This commit adds the Pause icon to the list of icons in the components/icons/index.tsx file. It also includes the retrainKbData and deleteKnowledgeBase functions in the http/requests.ts file. These changes enhance the functionality of the application by providing the ability to pause and retrain knowledge base data, as well as delete knowledge bases.