refactor: Add URL parameter to addKbPayload interface
This commit adds the `url` parameter to the `addKbPayload` interface in the `knowledgebase.controller.ts` file. The `url` parameter is used when the `type` is set to "WEB_PAGES" and represents the URL of the webpage to be added to the knowledge base. This change enhances the functionality of the payload structure and ensures consistency in handling different types of knowledge base data.