refactor: Update kb_ids validation in linkKbSchema
This commit updates the validation for the `kb_ids` property in the `linkKbSchema` object in the `schema_validation.ts` file. The previous implementation used the `string` type, but it has been changed to an array of strings using the `zod.array` method. This change ensures that multiple knowledge base IDs can be provided when linking a knowledge base, improving the flexibility and functionality of the schema.