refactor: Update match_embeddings function to accept kb_ids parameter
This commit updates the match_embeddings function in the function.sql file to accept a kb_ids parameter. The kb_ids parameter is an array of knowledge base IDs that the function will match against. This modification allows for more targeted similarity search by specifying the knowledge bases to search within. It improves the flexibility and functionality of the match_embeddings function.