feat: Add createConversationPayload type and update conversation service
This commit adds the `CreateConversationPayload` type to the `conversation.type.ts` file, which includes the `agent_id` property. It also updates the `createConversation` method in the `conversation.service.ts` file to use the `agent_id` and `userId` properties from the `data` parameter to create a new conversation in the database.