feat: Add method to fetch store information from API
This commit adds a new method `getStoreInfo()` to the `LS.service.ts` file in the `api` package. The method makes an API call to retrieve store information from the `https://api.lemonsqueezy.com/v1/stores` endpoint. It includes the necessary authorization header with the API key. The response data is then processed to extract the store IDs and names, which are returned as an array of objects.