feat: Update LS store IDs and Twilio product/variant IDs
This commit updates the LS store IDs and Twilio product/variant IDs in the code. The LS store ID in the `.env.example` file is changed to "95112" and a comment is added to indicate where to get it from. The Twilio product/variant IDs in the `sub-plan.ts` file are updated for both test and prod modes. The test IDs are changed to "318984" for the product and "463694" for the variant. The prod IDs are changed to "318978" for the product and "463687" for the variant.